refactor(web): move AgentSessionWrapper into pi-backend runtime
